Types_of_RAG_implementation_using_LangGraph/agentic-RAG/retriever.py
from langchain_chroma import Chroma
from langchain_openai import OpenAIEmbeddings
from dotenv import load_dotenv
load_dotenv()

vectorstore = Chroma(
    collection_name="rag-chroma",
    embedding_function=OpenAIEmbeddings(),
    persist_directory="./chroma_db",
)

retriever = vectorstore.as_retriever(k = 2, search_type="similarity")